+++ NEWS.md
@@ -1,3 +1,17 @@
+# palaeoverse 1.3.1
+
+* Fixed point plotting bug in tax_range_strat
+
+# palaeoverse 1.3.0
+
+* Added plot customisability to tax_range_time (#99)
+* Update palaeorotate to handle updates with GPlates API (#112)
+* New function to create stratigraphic range plots, tax_range_strat (#97)
+* Added age arguments (max_ma/min_ma) to bin_time (#106)
+* Added warning to bin_time (#104)
+* Improved link accessibility (#88)
+* Fix superfluous columns in palaeorotate (#110)
+
 # palaeoverse 1.2.1
 
 * Fixed handling of multiple models in palaeorotate (#92)